Refactor: 캐릭터 타입 삭제
