+++ src/app/id-card/[id]/page.tsx
@@ -2,7 +2,7 @@ import TopNavigation from '@/components/TopNavigation/TopNavigation';
 import idCardApi from '@/lib/api/domain/id-cards.api';
 import { Intro } from '@/modules/IdCardDetail/Intro';
 import { KeywordContentCard } from '@/modules/IdCardDetail/KeywordContentCard/KeywordContentCard.server';
-import { CharacterType } from '@/types/user.type';
+import { CharacterType } from '@/types/id-cards.type';
 
 const bgColors: Record<CharacterType, string> = {
   BUDDY: 'bg-buddy-100',
