--- src/types/user.type.ts
@@ -1,1 +1,0 @@
-export type CharacterType = 'TRUE' | 'PIPI' | 'TOBBY' | 'BUDDY';
